requirements.txt
flask==3.1.0
google-generativeai==0.4.0
SpeechRecognition==3.10.1
pyttsx3==2.90
sounddevice==0.5.1
Werkzeug==3.1.3
gunicorn==23.0.0